requirements.txt
grpcio
grpcio-tools
pika
docker